+++ Data/SBV/Examples/Queries/FourFours.hs
@@ -19,15 +19,17 @@
 -- and ask the SMT solver to find the appropriate fillings.
 -----------------------------------------------------------------------------
 
-{-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE DeriveAnyClass     #-}
 {-# LANGUAGE FlexibleInstances  #-}
+{-# LANGUAGE TemplateHaskell     #-}
+{-# LANGUAGE StandaloneDeriving  #-}
+{-# LANGUAGE DeriveDataTypeable  #-}
+{-# LANGUAGE DeriveAnyClass      #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
 module Data.SBV.Examples.Queries.FourFours where
 
 import Data.SBV
 import Data.SBV.Control
-import Data.Generics
 
 import Data.List (inits, tails)
 import Data.Maybe
@@ -36,12 +38,12 @@ import Data.Maybe
 -- and exponentiation will only be to the power @0@. This does restrict the search space, but is sufficient to
 -- solve all the instances.
 data BinOp = Plus | Minus | Times | Divide | Expt
-           deriving (Eq, Ord, Show, Read, Data, SymWord, HasKind, SMTValue)
+mkSymbolicEnumeration ''BinOp
 
 -- | Supported unary operators. Similar to 'BinOp' case, we will restrict square-root and factorial to
 -- be only applied to the value @4.
 data UnOp  = Negate | Sqrt | Factorial
-           deriving (Eq, Ord, Show, Read, Data, SymWord, HasKind, SMTValue)
+mkSymbolicEnumeration ''UnOp
 
 -- | Symbolic variant of 'BinOp'.
 type SBinOp = SBV BinOp